tools/sam_pdf/requirements.txt
PyMuPDF>=1.24.0
numpy>=1.24.0,<2.0
Pillow>=10.0.0

# Latest PaddleOCR stack
paddlepaddle>=3.2.0,<3.3.0
paddleocr>=3.4.0
